+++ MtgCollectionTracker/DesktopApp/DesktopApp.csproj
@@ -15,8 +15,4 @@
     <ProjectReference Include="..\DataAccess\DataAccess.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="MVVM\View\" />
-  </ItemGroup>
-
 </Project>
