+++ MtgCollectionTracker/DesktopApp/MVVM/Controls/SelectedCardControl.xaml
@@ -5,7 +5,7 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:local="clr-namespace:DesktopApp.MVVM.Controls"
              xmlns:viewmodel="clr-namespace:DesktopApp.MVVM.ViewModel"
-             d:DataContext="{d:DesignInstance Type=viewmodel:MainViewModel}"
+             d:DataContext="{d:DesignInstance Type=viewmodel:SelectedCardPrintViewModel}"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
     <Grid>
