+++ MtgCollectionTracker/DesktopApp/MVVM/View/SelectedCardPrintView.xaml
@@ -1,0 +1,20 @@
+﻿<Page x:Class="DesktopApp.MVVM.View.SelectedCardPrintView"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:DesktopApp.MVVM.View"
+      xmlns:controls="clr-namespace:DesktopApp.MVVM.Controls"
+      xmlns:viewmodel="clr-namespace:DesktopApp.MVVM.ViewModel"
+      mc:Ignorable="d" 
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="SelectedCardPrintView">
+
+    <Page.DataContext>
+        <viewmodel:SelectedCardPrintViewModel />
+    </Page.DataContext>
+
+    <Grid>
+        <controls:SelectedCardControl />
+    </Grid>
+</Page>
