+++ MtgCollectionTracker/DesktopApp/MVVM/ViewModel/SelectedCardPrintViewModel.cs
@@ -1,0 +1,19 @@
+﻿using DesktopApp.MVVM.Model;
+
+using Prism.Mvvm;
+
+namespace DesktopApp.MVVM.ViewModel
+{
+    internal class SelectedCardPrintViewModel : BindableBase
+    {
+        /// <summary>
+        /// The selected card to show.
+        /// </summary>
+        private CardPrint _selectedCardPrint;
+        public CardPrint SelectedCardPrint
+        {
+            get { return _selectedCardPrint; }
+            set { SetProperty(ref _selectedCardPrint, value); }
+        }
+    }
+}
